Refactor v2 validator steps
